Utilities/AppCenter/requirements.txt
PyYAML==6.0
requests==2.27.1
